Move custom matchers to includable file
Move the custom matchers to a matchers.hpp to enable re-use across test
units.
