+++ lib/test/segy.cpp
@@ -5,6 +5,7 @@
 #include <vector>
 
 #include <catch/catch.hpp>
+#include "matchers.hpp"
 
 #include <segyio/segy.h>
 #include <segyio/util.h>
@@ -22,30 +23,6 @@ std::string str( const slice& s ) {
            "," + std::to_string( s.step ) +
            ")";
 }
-
-class ApproxRange : public Catch::MatcherBase< std::vector< float > > {
-    public:
-        explicit ApproxRange( const std::vector< float >& xs ) :
-            lhs( xs )
-        {}
-
-        virtual bool match( const std::vector< float >& xs ) const override {
-            if( xs.size() != lhs.size() ) return false;
-
-            for( size_t i = 0; i < xs.size(); ++i )
-                if( xs[ i ] != Approx(this->lhs[ i ]) ) return false;
-
-            return true;
-        }
-
-        virtual std::string describe() const override {
-            using str = Catch::StringMaker< std::vector< float > >;
-            return "~= " + str::convert( this->lhs );
-        }
-
-    private:
-        std::vector< float > lhs;
-};
 
 struct Err {
     // cppcheck-suppress noExplicitConstructor
